+++ packages/browser/test/index.test.ts
@@ -151,4 +151,11 @@ describe('browser notifier', () => {
       done()
     })
   })
+
+  it('indicates whether or not the client is started', () => {
+    const Bugsnag = getBugsnag()
+    expect(Bugsnag.isStarted()).toBe(false)
+    Bugsnag.start(API_KEY)
+    expect(Bugsnag.isStarted()).toBe(true)
+  })
 })